+++ src/libms/ft_calloc.c
@@ -1,26 +1,28 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ms_has_opt.c                                       :+:      :+:    :+:   */
+/*   ft_calloc.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: tookuyam <tookuyam@student.42tokyo.fr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/11/16 01:51:43 by tookuyam          #+#    #+#             */
-/*   Updated: 2024/11/16 01:51:43 by tookuyam         ###   ########.fr       */
+/*   Created: 2024/11/16 02:59:09 by tookuyam          #+#    #+#             */
+/*   Updated: 2024/11/16 02:59:09 by tookuyam         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "ms_int_echo.h"
-#include <stdbool.h>
-#include <stddef.h>
+#include "libms.h"
+#include <stdint.h>
+#include <stdlib.h>
 
-bool	ms_has_opt(char *const argv[], const char *opt)
+void	*ft_calloc(size_t nmemb, size_t size)
 {
-	while (*argv != NULL)
-	{
-		if (ft_strcmp(*argv, opt) == 0)
-			return (true);
-		argv++;
-	}
-	return (false);
+	void	*ret;
+
+	if (size != 0 && nmemb > SIZE_MAX / size)
+		return (NULL);
+	ret = malloc(nmemb * size);
+	if (ret == NULL)
+		return (NULL);
+	ft_bzero(ret, nmemb * size);
+	return (ret);
 }
