+++ src/libms/ft_putstr_fd.c
@@ -1,29 +1,19 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   strncat.c                                          :+:      :+:    :+:   */
+/*   ft_putstr_fd.c                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: tookuyam <tookuyam@student.42tokyo.fr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/11/16 02:55:49 by tookuyam          #+#    #+#             */
-/*   Updated: 2024/11/16 02:55:49 by tookuyam         ###   ########.fr       */
+/*   Created: 2024/11/16 03:16:14 by tookuyam          #+#    #+#             */
+/*   Updated: 2024/11/16 03:16:14 by tookuyam         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "ms_int_echo.h"
-#include <stddef.h>
+#include "libms.h"
+#include <unistd.h>
 
-char	*ft_strncat(char *dest, const char *src, size_t n)
+void	ft_putstr_fd(char *s, int fd)
 {
-	size_t dest_len = ft_strlen(dest);
-	size_t i;
-
-	i = 0;
-	while(i < n && src[i] != '\0')
-	{
-		dest[dest_len + i] = src[i];
-		i++;
-	}
-	dest[dest_len + i] = '\0';
-	return (dest);
+	write(fd, s, ft_strlen(s));
 }
