+++ src/libms/ft_strlen.c
