+++ src/libms/ft_strncat.c
@@ -1,28 +1,29 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_calloc.c                                        :+:      :+:    :+:   */
+/*   strncat.c                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: tookuyam <tookuyam@student.42tokyo.fr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/11/16 02:59:09 by tookuyam          #+#    #+#             */
-/*   Updated: 2024/11/16 02:59:09 by tookuyam         ###   ########.fr       */
+/*   Created: 2024/11/16 02:55:49 by tookuyam          #+#    #+#             */
+/*   Updated: 2024/11/16 02:55:49 by tookuyam         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "ms_int_echo.h"
-#include <stdint.h>
-#include <stdlib.h>
+#include "libms.h"
+#include <stddef.h>
 
-void	*ft_calloc(size_t nmemb, size_t size)
+char	*ft_strncat(char *dest, const char *src, size_t n)
 {
-	void	*ret;
+	size_t dest_len = ft_strlen(dest);
+	size_t i;
 
-	if (size != 0 && nmemb > SIZE_MAX / size)
-		return (NULL);
-	ret = malloc(nmemb * size);
-	if (ret == NULL)
-		return (NULL);
-	ft_bzero(ret, nmemb * size);
-	return (ret);
+	i = 0;
+	while(i < n && src[i] != '\0')
+	{
+		dest[dest_len + i] = src[i];
+		i++;
+	}
+	dest[dest_len + i] = '\0';
+	return (dest);
 }
